# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateProductRequest:
    properties:
      brand_name:
        type: string
      category_name:
        type: string
      description:
        type: string
      name:
        type: string
      variants:
        items:
          $ref: '#/definitions/handlers.VariantRequest'
        minItems: 1
        type: array
    required:
    - category_name
    - name
    - variants
    type: object
...
  handlers.UpdateProductRequest:
    properties:
      brand_name:
        type: string
      category_name:
        type: string
      description:
        type: string
      name:
        type: string
      variants:
        items:
          $ref: '#/definitions/handlers.VariantRequest'
        minItems: 1
        type: array
    required:
    - category_name
    - name
    - variants
    type: object
...
  handlers.VariantRequest:
    properties:
      color:
        type: string
      image:
        type: string
      price:
        minimum: 0
        type: number
      size:
        type: string
      stock:
        minimum: 0
        type: integer
    required:
    - color
    - price
    - size
    - stock
    type: object
...
  models.Product:
    properties:
      brand_id:
        type: string
      category_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_deleted:
        type: boolean
      max_price:
        type: number
      min_price:
        type: number
      name:
        type: string
      total_stock:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
      variants:
        items:
          $ref: '#/definitions/models.ProductVariant'
        type: array
    type: object
  models.ProductVariant:
    properties:
      color:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      image:
        type: string
      is_active:
        type: boolean
      is_deleted:
        type: boolean
      price:
        type: number
      product_id:
        type: string
      size:
        type: string
      stock:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
info:
  contact: {}
  description: A RESTful API for a clothes shop built with Golang and Gin.
  title: Clothes Shop API
  version: "1.0"
paths:
...
  /product-variants/{id}/soft-delete:
    delete:
      consumes:
      - application/json
      description: Mark a product variant as deleted (soft delete)
      parameters:
      - description: Product Variant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductVariant'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Soft delete a product variant
      tags:
      - product-variants
  /product-variants/{id}/toggle-active:
    patch:
      consumes:
      - application/json
      description: Toggle the active status of a product variant (activate/deactivate)
      parameters:
      - description: Product Variant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductVariant'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Toggle product variant active status
      tags:
      - product-variants
  /products:
    get:
      consumes:
      - application/json
      description: Retrieve a list of products with optional pagination, price filter,
        category filter, brand filter, and search
      parameters:
      - default: 1
        description: Page number (default 1)
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page (default 10)
        in: query
        name: limit
        type: integer
      - description: Minimum price filter
        in: query
        name: min_price
        type: number
      - description: Maximum price filter
        in: query
        name: max_price
        type: number
      - description: Category name filter
        in: query
        name: category
        type: string
      - description: Brand name filter
        in: query
        name: brand
        type: string
      - description: Product name search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      summary: Get all products with pagination and filters
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product with the provided details
      parameters:
      - description: Product creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new product
      tags:
      - products
  /products/{id}:
    put:
      consumes:
      - application/json
      description: Update a product with the provided details
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an existing product
      tags:
      - products
  /products/{id}/soft-delete:
    delete:
      consumes:
      - application/json
      description: Mark a product as deleted (soft delete)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Soft delete a product
      tags:
      - products
  /products/{id}/toggle-active:
    patch:
      consumes:
      - application/json
      description: Toggle the active status of a product (activate/deactivate)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Toggle product active status
      tags:
      - products
securityDefinitions:
  BearerAuth:
    description: Enter JWT token without "Bearer" prefix.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"